Re-organized action bar cards
