+++ src/components/battle-planner/action-bar/realms-card/RealmsCardListContext.ts
@@ -1,0 +1,8 @@
+import { EntityId } from '@reduxjs/toolkit';
+
+import createTabbedContext from '../../../common/TabbedContext';
+
+export const {
+  useContext: useRealmsCardListContext,
+  ContextProvider: RealmsCardListContextProvider,
+} = createTabbedContext<EntityId>();
